Allow positional arguments for pattern
