+++ main.go
@@ -1,12 +1,7 @@
 package main
 
 import (
-	"flag"
 	"fmt"
-	"os"
-	"runtime"
-	"strings"
-	"sync"
 )
 
 var IGNORE_PATHS = [7]string{
@@ -31,41 +26,20 @@ func printBuildInfo() {
 }
 
 func main() {
-	//The number of workers. If there are more workers the system can read from
-	//the work queue more often and a larger queue is not required.
-	workers := flag.Int("w", -1, "Number of workers")
-	//If the queue overflows we'll use a slice to store work which might slow the system
-	queueSize := flag.Int("q", 512, "The max work queue size")
-	maxResults := flag.Int("c", -1, "The maximum number of results to find")
-	dir := flag.String("d", ".", "The starting directory to check for files")
-	pattern := flag.String("p", "", "A pattern to check for within the file names")
-	v := flag.Bool("v", false, "print the version and build information")
-	flag.Parse()
-
-	if *v {
+	cfg := parseCLI()
+	if cfg.Version {
 		printBuildInfo()
 		return
 	}
 
-	if *pattern == "" {
-		fmt.Println("No pattern provided")
+	if cfg.Pattern == "" {
+		fmt.Println("No pattern found, please provide a search pattern")
 		return
 	}
 
-	w := *workers
-	if *workers <= 0 {
-		w = runtime.NumCPU() + 2
-	}
-
-	//Only for OSX/Linux, sorry windows
-	//Remove any trailing slashes in the path
-	if (*dir)[len(*dir)-1:] == "/" {
-		*dir = string((*dir)[0 : len(*dir)-1])
-	}
-
-	printCh := make(chan string, w)
+	printCh := make(chan string, cfg.Workers)
 	//The system will reach deadlock if the work queue reaches capacity
-	workQ := make(chan string, *queueSize)
+	workQ := make(chan string, cfg.QueueSize)
 	//To avoid deadlock, send tasks here which will have a non-blocky retry
 	//func to add tasks back to workQ
 	failover := make(chan string)
@@ -77,110 +51,11 @@ func main() {
 	//defer close(dirCount)
 	//defer close(failover)
 	go handleFailover(workQ, failover)
-	go createWorkerPool(pattern, workQ, failover, printCh, dirCount, w)
+	go createWorkerPool(cfg.Pattern, workQ, failover, printCh, dirCount, cfg.Workers)
 
 	//Send first work request
-	workQ <- *dir
+	workQ <- cfg.Dir
 
 	//Print all results
-	showResults(printCh, maxResults)
-}
-
-func showResults(ch chan string, limit *int) {
-	if *limit > 0 {
-		n := 0
-		for item := range ch {
-			fmt.Println(item)
-			n++
-			if n >= *limit {
-				return
-			}
-		}
-	} else {
-		for item := range ch {
-			fmt.Println(item)
-		}
-	}
-}
-
-func dirChecker(in chan int, work chan string) {
-	n := 1
-	for i := range in {
-		n += i
-		if n <= 0 {
-			close(work)
-			return
-		}
-	}
-}
-
-func createWorkerPool(p *string, in chan string, failover chan string, results chan string, cnt chan int, w int) {
-	var wg sync.WaitGroup
-	for i := 0; i < w; i++ {
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			search(p, in, failover, results, cnt)
-		}()
-	}
-	wg.Wait()
-	close(results)
-}
-func search(pattern *string, in chan string, failover chan string, out chan string, cnt chan int) {
-	for path := range in {
-		items, err := os.ReadDir(path)
-		if err != nil {
-			fmt.Println("Error reading the directory", path)
-			fmt.Println(err)
-			cnt <- -1
-		}
-	ItemSearch:
-		for _, item := range items {
-			if item.IsDir() {
-				//Don't dive into directories I don't care about
-				for _, p := range IGNORE_PATHS {
-					if p == item.Name() {
-						continue ItemSearch
-					}
-				}
-				subPath := fmt.Sprintf("%s/%s", path, item.Name())
-				cnt <- 1
-				select {
-				case in <- subPath:
-				case failover <- subPath:
-				}
-			}
-			//Always check if the name of the thing matches pattern, including directory names
-			if strings.Index(item.Name(), *pattern) >= 0 {
-				//subPath is repeated but no point in creating an allocation if not required
-				subPath := fmt.Sprintf("%s/%s", path, item.Name())
-				out <- subPath
-			}
-
-		}
-		//We finished reading everything in the dir, tell the accounted we finished
-		cnt <- -1
-	}
-}
-
-func handleFailover(work, fail chan string) {
-	var q []string
-	for {
-		task := <-fail
-		q = append(q, task)
-		//TODO: Add verbose logging here so users can check if the failover was used
-		for {
-			select {
-			case work <- q[0]:
-				q = q[1:]
-			case task := <-fail:
-				q = append(q, task)
-			default:
-			}
-			//I don't know if we'll get an issue with `work <- q[0]` unless we have this
-			if len(q) == 0 {
-				break
-			}
-		}
-	}
+	showResults(printCh, &cfg.MaxResults)
 }
